fix(support): harden WhatsApp webhook (XSS, log-injection) + move logic to a service
CodeQL + claude[bot] on #183:
- HIGH reflected XSS: the verify handshake echoed the raw hub.challenge. Echo it
  only when it's purely numeric (Meta always sends an integer) and as text/plain,
  so a crafted value can never be reflected as markup.
- MEDIUM log-injection: stop logging the user-controlled hub.mode.
- Package layout: move signature verification + payload parsing into
  WhatsAppInboundService(+Impl); the controller is now thin HTTP wiring only.
